requirements.txt
streamlit
sentence-transformers
opencv-python-headless
pillow
torch
numpy
protobuf